added auth
added auth

+++ backend/routes/posts.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const Post = require('../models/post');
+const checkAuth = require("../middleware/check-auth");
 const multer = require("multer");
 
 const router = express.Router();
@@ -26,12 +27,13 @@ const storage = multer.diskStorage({
   }
 });
 
-router.post("", multer({storage: storage}).single("image"), (req, res, next) => {
+router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
   const url = req.protocol + '://' + req.get("host");
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -46,23 +48,48 @@ router.post("", multer({storage: storage}).single("image"), (req, res, next) => 
   });
 });
 
-router.put("/:id", (req, res, next) => {
+router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
+  let imagePath = req.body.imagePath;
+  if(req.file){
+    const url = req.protocol + "://" + req.get("host");
+    imagePath = url + "/images/" + req.file.filename
+  }
   const post = new Post({
     _id: req.body.id,
     title: req.body.title,
-    content: req.body.content
+    content: req.body.content,
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    console.log(result);
-    res.status(200).json({message: "Update successfull!"});
+  console.log(post);
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if(result.nModified > 0){
+      res.status(200).json({message: "Update successfull!"});
+    } else {
+      res.status(401).json({message: "Not authorized!"});
+    }
   })
 });
 
-router.get('',(req, res, next) => {
-  Post.find().then(documents => {
+router.get("",(req, res, next) => {
+  const pageSize = +req.query.pagesize;
+  const currentPage = +req.query.page;
+  const postQuery = Post.find();
+  let fetchedPosts;
+  if(pageSize && currentPage){
+    postQuery
+    .skip(pageSize * (currentPage - 1))
+    .limit(pageSize);
+  }
+  postQuery.then(documents => {
+    fetchedPosts = documents;
+      return Post.count();
+    })
+    .then(count => {
       res.status(200).json({
-        message: 'Posts fetched successfully!',
-        posts: documents
+        message: "Posts fetched successfully!",
+        posts: fetchedPosts,
+        maxPosts: count
       });
     });
 });
@@ -77,10 +104,14 @@ router.get("/:id", (req, res, next) => {
   });
 });
 
-router.delete("/:id", (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
+router.delete("/:id", checkAuth, (req, res, next) => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
     console.log(result);
-  res.status(200).json({message: "Post deleted!"});
+    if(result.n > 0){
+      res.status(200).json({message: "Delete successfull!"});
+    } else {
+      res.status(401).json({message: "Not authorized!"});
+    }
   });
 });
